- mp4v2 binary now contains ppc, i386 and x86_64 architectures. - fixed a potential crash. - small white spaces cleanup.

+++ mp4v2 wrapper/MP42File.h
@@ -29,7 +29,7 @@
 @property(readonly) MP42Metadata    *metadata;
 
 - (id)   initWithExistingFile:(NSString *) path andDelegate:(id) del;
-- (int)  tracksCount;
+- (NSInteger)  tracksCount;
 - (id)   trackAtIndex:(NSUInteger) index;
 - (void) addTrack:(id) track;
 - (void) removeTrackAtIndex:(NSUInteger) index;
